Cleanup
* Don't use the java.com namespace

 * Change the packages and app id to a more compliant name and shamelessly use the #devfleet moniker (org.devfleet.android.eveminer)

* Cleanup gradle configs
