+++ .gitignore
@@ -1,0 +1,6 @@
+.idea/
+app/build/
+app/local.properties
+app/app.iml
+build/
+.gradle/
