+++ app/src/main/java/org/devfleet/android/minerhelper/SystemC.java
@@ -1,16 +1,16 @@
-package com.randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6.EVE_Mining_Calc;
+package org.devfleet.android.minerhelper;
 
-import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
 
-public class SystemCgas extends AppCompatActivity {
+public class SystemC extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_system_c);
         getFragmentManager().beginTransaction()
-                .replace(android.R.id.content, new SystemCgasFragment())
+                .replace(android.R.id.content, new SystemCFragment())
                 .commit();
     }
 }
