+++ app/src/main/java/org/devfleet/android/minerhelper/SystemCFragment.java
@@ -1,15 +1,15 @@
-package com.randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6.EVE_Mining_Calc;
+package org.devfleet.android.minerhelper;
 
-import android.app.Fragment;
 import android.os.Bundle;
 import android.preference.PreferenceFragment;
 
-public class SystemCgasFragment extends PreferenceFragment {
+
+public class SystemCFragment extends PreferenceFragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         // Load the preferences from an XML resource
-        addPreferencesFromResource(R.xml.systemprefsgas);
+        addPreferencesFromResource(R.xml.systemprefs);
     }
 }
