+++ app/src/main/java/org/devfleet/android/minerhelper/SystemCice.java
@@ -1,4 +1,4 @@
-package com.randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6.EVE_Mining_Calc;
+package org.devfleet.android.minerhelper;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
