+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:ads="http://schemas.android.com/apk/res-auto"
-    tools:context="com.randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6.EVE_Mining_Calc.MainActivity">
+    tools:context="org.devfleet.android.minerhelper.MainActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/RecV"
