+++ app/src/main/res/layout/activity_system_c.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6.EVE_Mining_Calc.SystemC">
+    tools:context="org.devfleet.android.minerhelper.SystemC">
 
 </android.support.constraint.ConstraintLayout>
