--- app/src/test/java/com/randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6/EVE_Mining_Calc/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package com.randomlettersandnumbers15645151gd65fg16d5fgs46r84tserg51f3d21ger6.EVE_Mining_Calc;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
